Reorganize in/out registrations tests in explorer
Split them in a subfolder, this will allow to add more tests on subject
like the discord formatting without overloading existing files.
